requirements.txt
Django==2.1.1
psycopg2
psycopg2-binary
django-crispy-forms
django-filter
